+++ extendr-api/src/wrapper/matrix.rs
@@ -1,7 +1,6 @@
 //! Wrappers for matrices with deferred arithmetic.
-
+use self::robj::{AsTypedSlice, Robj};
 use super::*;
-use crate::robj::GetSexp;
 use crate::scalar::Scalar;
 use std::ops::{Index, IndexMut};
 
@@ -31,6 +30,7 @@ pub struct RArray<T, D> {
 
     /// Dimensions of the array.
     dim: D,
+
     _data: std::marker::PhantomData<T>,
 }
 
@@ -66,7 +66,6 @@ where
     pub fn new_with_na(nrow: usize, ncol: usize) -> Self {
         let mut matrix = Self::new(nrow, ncol);
         if nrow != 0 || ncol != 0 {
-            // matrix.data_mut().iter_mut().for_each(|x| {
             matrix
                 .as_typed_slice_mut()
                 .unwrap()
@@ -125,10 +124,9 @@ impl<T> Offset<[usize; 3]> for RArray<T, [usize; 3]> {
     }
 }
 
-impl<'a, T, D> RArray<T, D>
+impl<T, D> RArray<T, D>
 where
-    T: 'a,
-    Robj: AsTypedSlice<'a, T>,
+    Robj: for<'a> AsTypedSlice<'a, T>,
 {
     pub fn from_parts(robj: Robj, dim: D) -> Self {
         Self {
@@ -139,12 +137,12 @@ where
     }
 
     /// Returns a flat representation of the array in col-major.
-    pub fn data(&self) -> &'a [T] {
+    pub fn data(&self) -> &[T] {
         self.as_typed_slice().unwrap()
     }
 
-    /// Returns a flat mutable representation of the array in col-major.
-    pub fn data_mut(&mut self) -> &'a mut [T] {
+    /// Returns a flat, mutable representation of the array in col-major.
+    pub fn data_mut(&mut self) -> &mut [T] {
         self.as_typed_slice_mut().unwrap()
     }
 
@@ -154,9 +152,10 @@ where
     }
 }
 
-impl<'a, T: ToVectorValue + 'a> RColumn<T>
+impl<T> RColumn<T>
 where
-    Robj: AsTypedSlice<'a, T>,
+    T: ToVectorValue,
+    Robj: for<'a> AsTypedSlice<'a, T>,
 {
     /// Make a new column type.
     pub fn new_column<F: FnMut(usize) -> T>(nrows: usize, f: F) -> Self {
@@ -172,9 +171,10 @@ where
     }
 }
 
-impl<'a, T: ToVectorValue + 'a> RMatrix<T>
+impl<T> RMatrix<T>
 where
-    Robj: AsTypedSlice<'a, T>,
+    T: ToVectorValue,
+    Robj: for<'a> AsTypedSlice<'a, T>,
 {
     /// Create a new matrix wrapper.
     ///
@@ -214,9 +214,10 @@ where
     }
 }
 
-impl<'a, T: ToVectorValue + 'a> RMatrix3D<T>
+impl<T> RMatrix3D<T>
 where
-    Robj: AsTypedSlice<'a, T>,
+    T: ToVectorValue,
+    Robj: for<'a> AsTypedSlice<'a, T>,
 {
     pub fn new_matrix3d<F: Clone + FnMut(usize, usize, usize) -> T>(
         nrows: usize,
@@ -254,71 +255,120 @@ where
     }
 }
 
-impl<'a, T: 'a> TryFrom<Robj> for RColumn<T>
+impl<T> TryFrom<&Robj> for RColumn<T>
 where
-    Robj: AsTypedSlice<'a, T>,
+    Robj: for<'a> AsTypedSlice<'a, T>,
 {
     type Error = Error;
 
-    fn try_from(mut robj: Robj) -> Result<Self> {
-        if let Some(_slice) = robj.as_typed_slice_mut() {
+    fn try_from(robj: &Robj) -> Result<Self> {
+        if let Some(_slice) = robj.as_typed_slice() {
             let len = robj.len();
-            Ok(RArray::from_parts(robj, [len]))
+            Ok(RArray::from_parts(robj.clone(), [len]))
         } else {
-            Err(Error::ExpectedVector(robj))
+            Err(Error::ExpectedVector(robj.clone()))
         }
     }
 }
 
-impl<'a, T: 'a> TryFrom<Robj> for RMatrix<T>
+impl<T> TryFrom<&Robj> for RMatrix<T>
 where
-    Robj: AsTypedSlice<'a, T>,
+    Robj: for<'a> AsTypedSlice<'a, T>,
 {
     type Error = Error;
 
-    fn try_from(mut robj: Robj) -> Result<Self> {
+    fn try_from(robj: &Robj) -> Result<Self> {
         if !robj.is_matrix() {
-            Err(Error::ExpectedMatrix(robj))
-        } else if let Some(_slice) = robj.as_typed_slice_mut() {
+            Err(Error::ExpectedMatrix(robj.clone()))
+        } else if let Some(_slice) = robj.as_typed_slice() {
             if let Some(dim) = robj.dim() {
                 let dim: Vec<_> = dim.iter().map(|d| d.inner() as usize).collect();
                 if dim.len() != 2 {
-                    Err(Error::ExpectedMatrix(robj))
+                    Err(Error::ExpectedMatrix(robj.clone()))
                 } else {
-                    Ok(RArray::from_parts(robj, [dim[0], dim[1]]))
+                    Ok(RArray::from_parts(robj.clone(), [dim[0], dim[1]]))
                 }
             } else {
-                Err(Error::ExpectedMatrix(robj))
+                Err(Error::ExpectedMatrix(robj.clone()))
             }
         } else {
-            Err(Error::TypeMismatch(robj))
+            Err(Error::TypeMismatch(robj.clone()))
         }
     }
 }
 
-impl<'a, T: 'a> TryFrom<Robj> for RMatrix3D<T>
+impl<T> TryFrom<&Robj> for RMatrix3D<T>
 where
-    Robj: AsTypedSlice<'a, T>,
+    Robj: for<'a> AsTypedSlice<'a, T>,
 {
     type Error = Error;
 
-    fn try_from(mut robj: Robj) -> Result<Self> {
-        if let Some(_slice) = robj.as_typed_slice_mut() {
+    fn try_from(robj: &Robj) -> Result<Self> {
+        if let Some(_slice) = robj.as_typed_slice() {
             if let Some(dim) = robj.dim() {
                 if dim.len() != 3 {
-                    Err(Error::ExpectedMatrix3D(robj))
+                    Err(Error::ExpectedMatrix3D(robj.clone()))
                 } else {
                     let dim: Vec<_> = dim.iter().map(|d| d.inner() as usize).collect();
-                    Ok(RArray::from_parts(robj, [dim[0], dim[1], dim[2]]))
+                    Ok(RArray::from_parts(robj.clone(), [dim[0], dim[1], dim[2]]))
                 }
             } else {
-                Err(Error::ExpectedMatrix3D(robj))
+                Err(Error::ExpectedMatrix3D(robj.clone()))
             }
         } else {
-            Err(Error::TypeMismatch(robj))
+            Err(Error::TypeMismatch(robj.clone()))
         }
     }
 }
+
+macro_rules! impl_try_from_robj_ref {
+    ($($type : tt)*) => {
+        $(
+            impl<T> TryFrom<Robj> for $type<T>
+            where
+                Robj: for<'a> AsTypedSlice<'a, T>,
+            {
+                type Error = Error;
+
+                fn try_from(robj: Robj) -> Result<Self> {
+                    <$type<T>>::try_from(&robj)
+                }
+            }
+
+            impl<T> TryFrom<&Robj> for Option<$type<T>>
+            where
+                Robj: for<'a> AsTypedSlice<'a, T>,
+            {
+                type Error = Error;
+
+                fn try_from(robj: &Robj) -> Result<Self> {
+                    if robj.is_null() || robj.is_na() {
+                        Ok(None)
+                    } else {
+                        Ok(Some(<$type<T>>::try_from(robj)?))
+                    }
+                }
+            }
+
+            impl<T> TryFrom<Robj> for Option<$type<T>>
+            where
+                Robj: for<'a> AsTypedSlice<'a, T>,
+            {
+                type Error = Error;
+
+                fn try_from(robj: Robj) -> Result<Self> {
+                    <Option::<$type<T>>>::try_from(&robj)
+                }
+            }
+        )*
+    }
+}
+
+impl_try_from_robj_ref!(
+    RMatrix
+    RColumn
+    RMatrix3D
+);
 
 impl<T, D> From<RArray<T, D>> for Robj {
     /// Convert a column, matrix or matrix3d to an Robj.
@@ -328,34 +378,33 @@ impl<T, D> From<RArray<T, D>> for Robj {
 }
 
 pub trait MatrixConversions: GetSexp {
-    fn as_column<'a, E: 'a>(&self) -> Option<RColumn<E>>
+    fn as_column<E>(&self) -> Option<RColumn<E>>
     where
-        Robj: AsTypedSlice<'a, E>,
+        Robj: for<'a> AsTypedSlice<'a, E>,
     {
-        <RColumn<E>>::try_from(self.as_robj().clone()).ok()
+        <RColumn<E>>::try_from(self.as_robj()).ok()
     }
 
-    fn as_matrix<'a, E: 'a>(&self) -> Option<RMatrix<E>>
+    fn as_matrix<E>(&self) -> Option<RMatrix<E>>
     where
-        Robj: AsTypedSlice<'a, E>,
+        Robj: for<'a> AsTypedSlice<'a, E>,
     {
-        <RMatrix<E>>::try_from(self.as_robj().clone()).ok()
+        <RMatrix<E>>::try_from(self.as_robj()).ok()
     }
 
-    fn as_matrix3d<'a, E: 'a>(&self) -> Option<RMatrix3D<E>>
+    fn as_matrix3d<E>(&self) -> Option<RMatrix3D<E>>
     where
-        Robj: AsTypedSlice<'a, E>,
+        Robj: for<'a> AsTypedSlice<'a, E>,
     {
-        <RMatrix3D<E>>::try_from(self.as_robj().clone()).ok()
+        <RMatrix3D<E>>::try_from(self.as_robj()).ok()
     }
 }
 
 impl MatrixConversions for Robj {}
 
-impl<'a, T> Index<[usize; 2]> for RArray<T, [usize; 2]>
+impl<T> Index<[usize; 2]> for RArray<T, [usize; 2]>
 where
-    T: 'a,
-    robj::Robj: robj::AsTypedSlice<'a, T>,
+    Robj: for<'a> AsTypedSlice<'a, T>,
 {
     type Output = T;
 
@@ -384,10 +433,9 @@ where
     }
 }
 
-impl<'a, T> IndexMut<[usize; 2]> for RArray<T, [usize; 2]>
+impl<T> IndexMut<[usize; 2]> for RArray<T, [usize; 2]>
 where
-    T: 'a,
-    robj::Robj: robj::AsTypedSlice<'a, T>,
+    Robj: for<'a> AsTypedSlice<'a, T>,
 {
     /// Zero-based mutable indexing in row, column order.
     ///
@@ -430,18 +478,17 @@ impl<T, D> DerefMut for RArray<T, D> {
 
 #[cfg(test)]
 mod tests {
+    use super::*;
+    use crate as extendr_api;
     use extendr_engine::with_r;
     use prelude::{Rcplx, Rfloat, Rint};
 
-    use super::*;
-
     #[test]
     fn test_empty_matrix_new() {
-        dbg!("print like R");
         with_r(|| {
             // These are arbitrarily filled. We cannot create assertions for them.
-            let m: RMatrix<Rbyte> = RMatrix::new(5, 2); // possible!
-            unsafe { Rf_PrintValue(m.get()) };
+            // let m: RMatrix<Rbyte> = RMatrix::new(5, 2); //   Error: Error: unimplemented type 'char' in 'eval'
+            // unsafe { Rf_PrintValue(m.get()) };
             let m: RMatrix<Rbool> = RMatrix::new(5, 2);
             unsafe { Rf_PrintValue(m.get()) };
             let m: RMatrix<Rint> = RMatrix::new(5, 2);
@@ -450,6 +497,7 @@ mod tests {
             unsafe { Rf_PrintValue(m.get()) };
             let m: RMatrix<Rcplx> = RMatrix::new(5, 2);
             unsafe { Rf_PrintValue(m.get()) };
+            rprintln!();
 
             // let m: RMatrix<Rbyte> = RMatrix::new_with_na(10, 2); // not possible!
             // unsafe { Rf_PrintValue(m.get()) };