+++ extendr-api/tests/from_iterator_tests.rs
@@ -63,7 +63,7 @@ fn test_with_gc_torture_strings() {
         let qq_r_character_vec: Strings = question_quote.into_iter().collect();
         R!("gctorture(on = FALSE)")?;
         // Strings::from_values is the same as `.collect`.
-        let qq_directly = Strings::from_robj(&Robj::from(question_quote)).unwrap();
+        let qq_directly = Strings::try_from(&Robj::from(question_quote)).unwrap();
         assert_eq!(qq_r_character_vec, qq_directly);
     );
 }