+++ tests/extendrtests/src/rust/src/lib.rs
@@ -21,7 +21,7 @@ fn hello_world() -> &'static str {
 #[extendr]
 fn do_nothing() {}
 
-// From: input/output conversion
+// TryFrom: conversions
 
 // Atomic types
 
@@ -68,60 +68,50 @@ fn double_vec(x: Vec<f64>) -> Vec<f64> {
 }
 
 // NA-related tests
-#[extendr(use_try_from = true)]
+#[extendr]
 fn try_rfloat_na() -> Rfloat {
     Rfloat::na()
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn try_rint_na() -> Rint {
     Rint::na()
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn check_rfloat_na(x: Rfloat) -> bool {
     x.is_na()
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn check_rint_na(x: Rint) -> bool {
     x.is_na()
 }
 
 // Non-atomic types
-// TODO
-
-// TryFrom: conversions
-
+// TOD
 // Atomic types
-
-// Convert a vector of doubles to itself
-// x a vector of doubles
-#[extendr(use_try_from = true)]
-fn try_double_vec(x: Vec<f64>) -> Vec<f64> {
-    x
-}
 
 // Non-atomic types
 // TODO
 
 // Vector wrappers
-#[extendr(use_try_from = true)]
+#[extendr]
 fn get_doubles_element(x: Doubles, i: i32) -> Rfloat {
     x.elt(i as usize)
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn get_integers_element(x: Integers, i: i32) -> Rint {
     x.elt(i as usize)
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn get_logicals_element(x: Logicals, i: i32) -> Rbool {
     x.elt(i as usize)
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn doubles_square(input: Doubles) -> Doubles {
     let mut result = Doubles::new(input.len());
 
@@ -132,7 +122,7 @@ fn doubles_square(input: Doubles) -> Doubles {
     result
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn complexes_square(input: Complexes) -> Complexes {
     let mut result = Complexes::new(input.len());
 
@@ -143,7 +133,7 @@ fn complexes_square(input: Complexes) -> Complexes {
     result
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn integers_square(input: Integers) -> Integers {
     let mut result = Integers::new(input.len());
 
@@ -154,7 +144,7 @@ fn integers_square(input: Integers) -> Integers {
     result
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn logicals_not(input: Logicals) -> Logicals {
     let mut result = Logicals::new(input.len());
 
@@ -167,7 +157,7 @@ fn logicals_not(input: Logicals) -> Logicals {
 
 // Parsing
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn check_default(#[default = "NULL"] x: Robj) -> bool {
     x.is_null()
 }
@@ -191,11 +181,7 @@ fn special_param_names(_x: i32, _y: i32) -> i32 {
 #[allow(non_snake_case)]
 fn __00__special_function_name() {}
 
-#[extendr(
-    use_try_from = true,
-    r_name = "test.rename.rlike",
-    mod_name = "test_rename_mymod"
-)]
+#[extendr(r_name = "test.rename.rlike", mod_name = "test_rename_mymod")]
 fn test_rename() -> i32 {
     1
 }
@@ -205,7 +191,7 @@ fn get_default_value(#[default = "42"] x: i32) -> i32 {
     x
 }
 
-#[extendr(use_try_from = true)]
+#[extendr]
 fn add_5_if_not_null(x: Nullable<Rint>) -> Nullable<Rint> {
     x.map(|y| y + 5)
 }
@@ -257,53 +243,6 @@ impl MyClass {
         x
     }
 }
-// Class for testing
-#[derive(Default, Debug)]
-struct MyClassTryFrom {
-    a: i32,
-}
-
-/// Class for testing (exported)
-/// @examples
-/// x <- MyClass$new()
-/// x$a()
-/// x$set_a(10)
-/// x$a()
-/// @export
-#[extendr(use_try_from = true)]
-impl MyClassTryFrom {
-    /// Method for making a new object.
-    fn new() -> Self {
-        Self { a: 0 }
-    }
-
-    /// Method for setting stuff.
-    /// @param x a number
-    fn set_a(&mut self, x: i32) {
-        self.a = x;
-    }
-
-    /// Method for getting stuff.
-    fn a(&self) -> i32 {
-        self.a
-    }
-
-    /// Method for getting one's self.
-    fn me(&self) -> &Self {
-        self
-    }
-
-    // https://github.com/extendr/extendr/issues/431
-    fn restore_from_robj(robj: Robj) -> Self {
-        let res: ExternalPtr<MyClassTryFrom> = robj.try_into().unwrap();
-        Self { a: res.a }
-    }
-
-    // https://github.com/extendr/extendr/issues/435
-    fn get_default_value(#[default = "42"] x: i32) -> i32 {
-        x
-    }
-}
 
 // Class for testing special names
 #[derive(Default, Debug)]
@@ -321,22 +260,6 @@ impl __MyClass {
     fn __name_test(&self) {}
 }
 
-// Class for testing special names
-#[derive(Default, Debug)]
-struct __MyClassTryFrom {}
-
-// Class for testing special names
-// Unexported because of documentation conflict
-#[extendr(use_try_from = true)]
-impl __MyClassTryFrom {
-    /// Method for making a new object.
-    fn new() -> Self {
-        Self {}
-    }
-    /// Method with special name unsupported by R
-    fn __name_test(&self) {}
-}
-
 // Class for testing (unexported)
 #[derive(Default, Debug)]
 struct MyClassUnexported {
@@ -346,26 +269,6 @@ struct MyClassUnexported {
 /// Class for testing (unexported)
 #[extendr]
 impl MyClassUnexported {
-    /// Method for making a new object.
-    fn new() -> Self {
-        Self { a: 22 }
-    }
-
-    /// Method for getting stuff.
-    fn a(&self) -> i32 {
-        self.a
-    }
-}
-
-// Class for testing (unexported)
-#[derive(Default, Debug)]
-struct MyClassUnexportedTryFrom {
-    a: i32,
-}
-
-/// Class for testing (unexported)
-#[extendr(use_try_from = true)]
-impl MyClassUnexportedTryFrom {
     /// Method for making a new object.
     fn new() -> Self {
         Self { a: 22 }
@@ -405,8 +308,6 @@ extendr_module! {
     fn char_vec;
     fn double_vec;
 
-    fn try_double_vec;
-
     fn get_doubles_element;
     fn get_integers_element;
     fn get_logicals_element;
@@ -437,10 +338,6 @@ extendr_module! {
     impl MyClass;
     impl __MyClass;
     impl MyClassUnexported;
-
-    impl MyClassTryFrom;
-    impl __MyClassTryFrom;
-    impl MyClassUnexportedTryFrom;
 
     fn my_device;